+++ source/extension.ts
@@ -1,26 +1,34 @@
 "use strict";
 import { commands, ExtensionContext, Uri } from "vscode";
+import * as path from "path";
 import * as fse from "fs-extra";
 
-import { getEngine, getFolderPath } from "./utils";
+import { getCurrentRoot, getEngine, getFolderPath } from "./utils";
 import { maybeRender, render } from "./render";
 import ask from "./ask-questions";
 import getTemplate from "./get-template";
 
 async function newFromTemplate(uri: Uri | undefined) {
-  const template = await getTemplate(uri);
+  const root = await getCurrentRoot(uri);
+
+  const template = await getTemplate(root);
   if (!template) return;
 
-  const engine = getEngine(uri);
+  const engine = getEngine(root);
   const answers = await ask(template.questions);
 
   template.files.forEach(({ name, content }) => {
     const folderName = maybeRender(engine, template.folder, answers);
-    const folderPath = getFolderPath(uri, folderName, template.defaultPath);
+    const folderPath = getFolderPath(
+      uri,
+      root,
+      folderName,
+      template.defaultPath,
+    );
     const fileName = render(engine, name, answers);
 
     fse.outputFileSync(
-      `${folderPath}/${fileName}`,
+      path.join(folderPath, fileName),
       render(engine, content.join("\n"), answers),
     );
   });